+++ lib/pay.rb
@@ -57,6 +57,9 @@ module Pay
   mattr_accessor :enabled_processors
   @@enabled_processors = [:stripe, :braintree, :paddle]
 
+  mattr_accessor :send_emails
+  @@send_emails = true
+
   mattr_accessor :emails
   @@emails = ActiveSupport::OrderedOptions.new
   @@emails.payment_action_required = true
@@ -111,12 +114,19 @@ module Pay
   end
 
   def self.send_email?(email_option, *remaining_args)
-    config_option = emails.send(email_option)
-
-    if config_option.respond_to?(:call)
-      config_option.call(*remaining_args)
+    if resolve_option(send_emails, *remaining_args)
+      email_config_option_enabled = emails.send(email_option)
+      resolve_option(email_config_option_enabled, *remaining_args)
     else
-      config_option
+      false
+    end
+  end
+
+  def self.resolve_option(option, *remaining_args)
+    if option.respond_to?(:call)
+      option.call(*remaining_args)
+    else
+      option
     end
   end
 end