+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+* Add new configuration option named `send_emails` which can be used to disable the sending of all current and future emails.
+  This option can be set to a boolean value or a proc/lambda that returns a boolean value. - @cjilbert504
+
 ### 6.4.0
 
 * Introduce `:pay` load hook - @excid3